Move datastores into their own package
