+++ lib/punt/datastore/clickhouse.go
@@ -1,4 +1,4 @@
-package punt
+package datastore
 
 import (
 	"database/sql"
