+++ lib/punt/datastore/datastore.go
@@ -1,4 +1,4 @@
-package punt
+package datastore
 
 import (
 	"log"
