+++ lib/punt/datastore/datastore_test.go
@@ -1,4 +1,4 @@
-package punt
+package datastore
 
 import (
 	"testing"
