+++ lib/punt/datastore/elasticsearch.go
@@ -1,4 +1,4 @@
-package punt
+package datastore
 
 import (
 	"context"
